fix: added a check for the presence of an image before deleting

+++ server/app/routes/images_bp.py
@@ -82,16 +82,18 @@ def add_image():
 @images_bp.route('/delete_image/<string:img_id>', methods=['DELETE'])
 def delete_image(img_id):
     image_info = db.images.find_one(ObjectId(img_id))
-    fs_id = image_info["fs_id"]
+    if (image_info):
+        fs_id = image_info["fs_id"]
 
-    db.images.delete_one({"_id": ObjectId(img_id)})
+        db.images.delete_one({"_id": ObjectId(img_id)})
 
-    map_fs.delete(fs_id)
-    for tile in tile_fs.find({"image_id": ObjectId(img_id)}):
-        tile_fs.delete(tile._id)
+        map_fs.delete(fs_id)
+        for tile in tile_fs.find({"image_id": ObjectId(img_id)}):
+            tile_fs.delete(tile._id)
 
-    socketio.emit("images", get_images_list())
-    return jsonify({'message': 'Image deleted successfully'})
+        socketio.emit("images", get_images_list())
+        return jsonify({'message': 'Image deleted successfully'})
+    return 'OK'
 
 
 # Маршрут для leaflet-а, возвращает кусочки для отображения.
